+++ src/main/java/com/ozay/repository/MemberRepository.java
@@ -119,7 +119,7 @@ public class MemberRepository {
             });
     }
 
-    public Member getMemberDetailByBuildingAndUserId(int id, int buildingId){
+    public Member getMemberDetailByBuildingAndUserId(long id, long buildingId){
         return (Member)jdbcTemplate.queryForObject("Select * FROM member WHERE building_id = ? AND user_id = ?",
             new Object[]{buildingId, id}, new MemberRowMapper(){
             });
