+++ src/main/java/com/ozay/web/rest/AccountResource.java
@@ -76,6 +76,7 @@ public class AccountResource {
     @Inject
     private UserBuildingRepository userBuildingRepository;
 
+
     /**
      * POST  /rest/register -> register the user.
      */
@@ -131,6 +132,7 @@ public class AccountResource {
             produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
     public ResponseEntity<String> activateAccount(@RequestParam(value = "key") String key) {
+
         return Optional.ofNullable(userService.activateRegistration(key))
             .map(user -> new ResponseEntity<String>(HttpStatus.OK))
             .orElse(new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR));
@@ -240,7 +242,7 @@ public class AccountResource {
         method = RequestMethod.GET,
         produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
-    public ResponseEntity<UserDTO> getAccountWithBuildingId(@PathVariable int buildingId) {
+    public ResponseEntity<UserDTO> getAccountWithBuildingId(@PathVariable Long buildingId) {
 
         return Optional.ofNullable(userService.getUserWithAuthorities(buildingId))
             .map(user -> new ResponseEntity<>(
